pyproject.toml
[project]
name = "sei-dashboard"
version = "0.1.0"
description = "Automação do SEI (download, QA e extração offline)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "playwright==1.45.0",
  "python-dotenv==1.0.1",
  "PySide6==6.7.1",
  "fastapi==0.114.0",
  "uvicorn[standard]==0.30.6",
  "SQLAlchemy==2.0.32",
  "passlib[bcrypt]==1.7.4",
  "python-jose[cryptography]==3.3.0",
  "pydantic[email]==2.8.2",
  "pdfplumber==0.11.4",
  "openpyxl==3.1.5",
  "beautifulsoup4==4.12.3",
  "python-dateutil==2.9.0.post0",
  "numpy>=1.24,<2",
  "rich==13.9.4",
  "sentencepiece==0.2.0",
  "transformers==4.38.2",
  "tokenizers==0.15.2",
  "torch==2.1.2",
]

[tool.uv]
package = false
